Moved addEvent to a component and created updateEvent
